requirements.txt
annotated-types==0.7.0
anthropic
anyio==4.12.1
appdirs
brotlicffi==1.2.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
datasets==3.2.0
distro==1.9.0
editdistance
exceptiongroup==1.3.1
filelock==3.20.2
fire
fonttools==4.61.1
fsspec==2025.12.0
google-generativeai
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
nltk
numpy==2.2.6
openai==2.14.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pip==25.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.0
pyparsing==3.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich
safetensors==0.7.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tempdir
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
tree_sitter<=0.21.3
tree_sitter_languages
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
vllm
wget
wheel==0.45.1
win_inet_pton==1.1.0
